Made a messagemanager util
All messages sent to players are now run through MessageManager

Problem though :) Misused some of the message methods.
When creating a faction through GUI, we shouldn't add the Faction prefix.
When a player joins the faction, all members receive a welcome message meant for just the joiner.
When kicked from a faction the kicked receives a message with the kicker name and faction name swapped.

Probably more unintended bugs.

TO BE FIXED